% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_diff.r
\name{auc_diff}
\alias{auc_diff}
\title{auc_diff function}
\usage{
auc_diff(dat, v1, v2, nv, kv)
}
\arguments{
\item{dat}{N by (M+1) matrix having variables in the order of cbind(y,x)}

\item{v1}{This can be set as v1=c(1) or v1=c(1,2)}

\item{v2}{This can be set as v2=c(2), v2=c(3), v2=c(1,3) or v2=c(3,4)}

\item{nv}{Sample size}

\item{kv}{Population prevalence}
}
\value{
This function will estimate significant difference between two PRS (either dependent or independent and joint or single). To get the test statistics for the difference between AUC(y~x[,v1]) and AUC(y~x[,v2]). (here we define AUC1=AUC(y~x[,v1])) and AUC2=AUC(y~x[,v2]))). The outputs are listed as follows.
\item{mean_diff}{AUC differences between AUC1 and AUC2}
\item{var}{Variances of AUC differences}
\item{upper_diff}{Upper value of the differences}
\item{lower_diff}{Upper value of the differences}
\item{p}{two tailed P-value for significant difference between AUC1 and AUC2}
\item{p_one_tail}{tailed P-value for significant difference}
\item{heller_p}{P-value based on Heller's test for significant difference}
\item{heller_upper_diff}{Upper limit of 95\% CI for the difference basedon Heller's test}
\item{heller_lower_diff}{Lower limit of 95\% CI for the difference basedon Heller's test}
}
\description{
This function estimates var(AUC(y~x[,v1]) - AUC(y~x[,v2]))
where AUC is the Area Under ROC curve of the model,
y is N by 1 matrix having the dependent variable, and
x is N by M matrix having M explanatory variables.
v1 or v2 indicates the ith column in the x matrix
(v1 or v2 can be multiple values between 1 - M, see Arguments below)
}
\keyword{R2}
\keyword{information}
\keyword{matrix}
\keyword{variance}
