% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat1.r
\docType{data}
\name{dat1}
\alias{dat1}
\title{Raw phenotypes and 2 sets of discovery PGSs}
\format{
A data frame with 10000 rows and 3 variables:
\describe{
\item{V1}{Phenotype, raw case-caontrol data}
\item{V2}{PGS1, for discovery population 1}
\item{V3}{PGS2, for discovery population 2}

}
}
\usage{
dat1
}
\description{
A dataset containing phenotypes and multiple PGSs estimated two independent discovery population
}
\keyword{datasets}
