% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat2.r
\docType{data}
\name{dat2}
\alias{dat2}
\title{Pre-adjusted phenotypes and 2 sets of discovery PGSs}
\format{
A data frame with 10000 rows and 3 variables:
\describe{
\item{V1}{Phenotype, preadjustde case-caontrol data}
\item{V2}{PGS1, for discovery population 1}
\item{V3}{PGS2, for discovery population 2}

}
}
\usage{
dat2
}
\description{
A dataset containing preadjusted phenotypes and multiple PGSs estimated two independent discovery population
}
\keyword{datasets}
