% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoa_auc12.r
\name{olkin_auc12}
\alias{olkin_auc12}
\title{olkin_auc12 function}
\usage{
olkin_auc12(omat, nv, kv)
}
\arguments{
\item{omat}{3 by 3 matrix having the correlation coefficients between y, x1 and x2, i.e. omat=cor(dat) where dat is N by 3 matrix having variables in the order of cbind (y,x1,x2)}

\item{nv}{Sample size}
}
\value{
This function will be used as source code
}
\description{
olkin_auc12 function
}
\keyword{source}
