% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoa_auc1_2.r
\name{olkin_auc1_2}
\alias{olkin_auc1_2}
\title{olkin_auc1_2 function}
\usage{
olkin_auc1_2(omat, nv, kv)
}
\arguments{
\item{omat}{3 by 3 matrix having the correlation coefficients between y, x1 and x2, i.e. omat=cor(dat) where dat is N by 3 matrix having variables in the order of cbind (y,x1,x2)}

\item{nv}{Sample size}

\item{kv}{Population prevalance}
}
\value{
This function will be used as source code
}
\description{
olkin_auc1_2 function
}
\keyword{source}
